Add Item functionality improvement
More functionalities will be added to the add item screen as well as attempts to improve the barcode scanning. Text recognition was implemented with the same code as the barcode reading and a box is drawn onscreen when barcode or text is recognised.
A list of smaller features was added:
1.	An remove all button on the list page with a check
2.	Correct removal of barcodes from LiveData variable
3.	Zoomed in view